Finalizando o funcionamento do projeto

+++ 04-ProjetoEstudante-nova-table-300923/02-ProjetoEstudante-nova-tabela/src/main/java/com/estudante/estudante/controller/AlunoController.java
@@ -1,11 +1,11 @@
 package com.estudante.estudante.controller;
 
-
 import com.estudante.estudante.model.Aluno;
 import com.estudante.estudante.repository.AlunoRepository;
 import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,79 +23,62 @@ public class AlunoController {
 
     // Como o professor mostrou, utilizando métodos do JpaRepo
     @GetMapping("/todos")
-    public ResponseEntity<List<Aluno>> findAll(){
+    public ResponseEntity<List<Aluno>> findAll() {
         return ResponseEntity.ok(repository.findAll());
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Aluno> GetById(@PathVariable long id){
+    public ResponseEntity<Aluno> GetById(@PathVariable long id) {
         return repository.findById(id)
                 .map(resp -> ResponseEntity.ok(resp))
                 .orElse(ResponseEntity.notFound().build());
     }
 
+
+
+
+
     // Método POST funciona ok
     // Método do professor
     @PostMapping("/alunos")
-    public ResponseEntity<Aluno> post (@RequestBody Aluno aluno){
-
-        //return ResponseEntity.status(HttpStatus.CREATED).body(REPOSITORY.save(aluno));
+    public ResponseEntity<Aluno> post(@RequestBody Aluno aluno) {
+        aluno.setAlunoId(0);
         return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(aluno));
-
     }
-
-    /**
-    // Original
-    @PostMapping("/alunos")
-    public Aluno adicionarAluno(@RequestBody Aluno aluno){
-        //aluno.setId(0);
-        // Forçando salvar um item novo, ao invés de update
-        Aluno alunoDb = repository.save(aluno);
-        return alunoDb;
-    }
-    **/
-
-
 
     // Método PUT N FUNCIONA como fazer sacoisa funcionar?
+    /*
 
-    @PutMapping("/alunos")
-    public ResponseEntity<Aluno> put(@RequestBody Aluno aluno){
+    public ResponseEntity<Aluno> put(@RequestBody Aluno aluno) {
         return ResponseEntity.status(HttpStatus.OK).body(repository.save(aluno));
     }
+    */
 
 
-    /*
+    @PutMapping("/alunos")
     public Aluno autalizarAluno(@RequestBody Aluno aluno){
         Aluno alunoDb = repository.save(aluno);
         return alunoDb;
     }
 
-     */
 
-    // Método DELETE funciona, soh preciso melhorar as mensagens confirmando se foi deletado ou n
+
+
+
+
+
+    // Método DELETE, com mensagem de retorno de confirmação/ID não encontrado
     @DeleteMapping("/alunos/{id}")
-    public void delete(@PathVariable long id){
-        repository.deleteById(id);
+    public String delete(@PathVariable long id) {
+        Optional<Aluno> temp = Optional.of(new Aluno());
+        temp = repository.findById(id);
 
-        //throw new RuntimeException("Id de aluno não encontrado: " + id);
-
-
-
-    }
-
-
-    /*
-    public String deletarAluno(@PathVariable long id){
-        Optional<Aluno> aluno = repository.findById(id);
-        // Jogando exceção se null
-        if (aluno == null){
-            throw new RuntimeException("Id de aluno não encontrado: " + id);
+        if (temp.isPresent()) {
+            repository.deleteById(id);
+            return "ID deletado: " + id;
+        } else {
+            return "Aluno não encontrado, ID: " + id;
         }
-        repository.deleteById(id);
-        return "Deletado aluno de id: " + id;
     }
-
-     */
 
 }
